% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avg_3dsk_info}
\alias{avg_3dsk_info}
\title{3DSK Average Ratings}
\format{
A data frame with 100 rows and 8 variables:
\describe{
\item{face_id}{the name ID from 3d.sk}
\item{face_gender}{face's gender (female, male)}
\item{face_age}{the face's age (in years)}
\item{att}{mean attractiveness rating}
\item{dom}{mean dominance rating}
\item{health}{mean health rating}
\item{sexdim}{mean sexual dimorphism rating; masculinity for male faces and femininity for female faces}
\item{trust}{mean trustworthiness rating}
}
}
\source{
\url{https://doi.org/10.17605/OSF.IO/A3947}
}
\usage{
avg_3dsk_info
}
\description{
A dataset containing information on the 100 images in the 3dsk image set \code{load_stim_3dsk()}. These images are from \href{https://www.3D.sk}{www.3D.sk}, purchased by Lisa DeBruine and Benedict Jones under their \href{https://www.3d.sk/static/commercial-terms}{commercial license}, which allows for sharing and publication of derivative images for educational use. Ratings are from 100 male and 100 female raters for each face_gender:trait on a scale from 1-7. See \code{raw_3dsk_info} for the unaggregated data and further details.
}
\examples{
head(avg_3dsk_info)
}
\keyword{datasets}
