% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_info}
\alias{canada_info}
\title{Canada Set Information}
\format{
A data frame with 40 rows and 3 variables:
\describe{
\item{face_id}{face's ID}
\item{face_gender}{face's gender (female, male)}
\item{face_age}{face's age (17-35 years)}
}
}
\source{
\url{https://figshare.com/articles/dataset/Young_Adult_White_Faces_with_Manipulated_Versions/4220517?file=7825783}
}
\usage{
canada_info
}
\description{
A dataset containing information on the 40 faces in the Young Adult White Faces with Manipulated Versions image set.
}
\keyword{datasets}
