% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stim.R
\name{load_stim}
\alias{load_stim}
\alias{load_stim_canada}
\alias{load_stim_london}
\alias{load_stim_neutral}
\alias{load_stim_zoom}
\alias{load_stim_smiling}
\alias{load_stim_composite}
\alias{load_stim_lisa}
\alias{load_stim_rainbow}
\alias{load_stim_3dsk}
\title{Load image set}
\usage{
load_stim(
  set = c("neutral_front", "smiling_front", "3dsk", "composite", "canada", "lisa",
    "zoom", "rainbow"),
  pattern = NULL
)

load_stim_canada(pattern = NULL)

load_stim_london(pattern = NULL)

load_stim_neutral(pattern = NULL)

load_stim_zoom(pattern = NULL)

load_stim_smiling(pattern = NULL)

load_stim_composite(pattern = NULL)

load_stim_lisa(pattern = NULL)

load_stim_rainbow(pattern = NULL)

load_stim_3dsk(pattern = NULL)
}
\arguments{
\item{set}{name of the image set}

\item{pattern}{Vector of patterns to use to search for files, or a vector of image indices}
}
\value{
a list with class stimlist
}
\description{
Load image sets as a stimlist, ready for processing with webmorphR functions. See Details below for citation links.
}
\details{
\itemize{
\item neutral/smiling: CC-BY \url{https://doi.org/10.6084/m9.figshare.5047666.v5}
\item composite: CC-BY \url{https://doi.org/10.6084/m9.figshare.4055130.v1}
\item canada: CC-BY \url{https://doi.org/10.6084/m9.figshare.4220517.v1}
\item 3dsk: \url{https://doi.org/10.17605/OSF.IO/A3947}, via \url{https://www.3d.sk} for educational use only
\item lisa/zoom: CC-0 images of the package creator
\item rainbow: see \code{rainbow_info} for individual image credits
}
}
\examples{
# loads the first 5 images in the set
stim <- load_stim("rainbow", 1:5)

# loads all images with 3 digits and m
canada_male <- load_stim("canada", "[0-9]{3}m")
}
