% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rainbow_info}
\alias{rainbow_info}
\title{Rainbow Set Information}
\format{
A data frame with 18 rows and 8 variables:
\describe{
\item{photo_name}{the name of the phto inthe simulus set}
\item{photo_id}{the numeric ID from Pixabay}
\item{photographer}{the photographer's name}
\item{user_name}{the photographer's pixabay username}
\item{user_id}{the photographer's pixabay user ID}
\item{colour}{the general colour of the image}
\item{type}{whether the image is an animal, landscape, or flower}
\item{url}{the pixabay URL for the image}
}
}
\usage{
rainbow_info
}
\description{
A dataset containing information on the 18 images in the rainbow image set. These images are from \href{https://pixabay.com}{Pixabay}, all using the \href{https://pixabay.com/service/license/}{Pixabay License}, so they are free for commercial and noncommercial use without attribution.
}
\keyword{datasets}
