% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_3dsk_info}
\alias{raw_3dsk_info}
\title{3DSK Raw Ratings}
\format{
A data frame with 100000 rows and 8 variables:
\describe{
\item{face_id}{the name ID from 3d.sk}
\item{face_gender}{face's gender (female, male)}
\item{face_age}{the face's age (in years)}
\item{rater_id}{the rater's anonymous ID}
\item{rater_gender}{rater's gender (female, male)}
\item{rater_age}{the rater's age (in years)}
\item{trait}{trait being rated (att, dom, health, sexdim, trust)}
\item{rating}{rating on a 1-7 scale}
}
}
\source{
\url{https://doi.org/10.17605/OSF.IO/A3947}
}
\usage{
raw_3dsk_info
}
\description{
A dataset containing information on the 100 images in the 3dsk image set \code{load_stim_3dsk()}. These images are from \href{https://www.3D.sk}{www.3D.sk}, purchased by Lisa DeBruine and Benedict Jones under their \href{https://www.3d.sk/static/commercial-terms}{commercial license}, which allows for sharing and publication of derivative images for educational use.
}
\details{
Face images were rated online at faceresearch.org. Raters were randomly allocated to rate either the 50 male or 50 female faces for one of attractiveness, dominance, health, sexual dimorphism (masculinity/femininity), or trustworthiness using 1 (much less \emph{trait adjective} than average) to 7 (much more \emph{trait adjective} than average). Male faces were rated for masculinity and female faces for femininity.

Data are from the first one hundred heterosexual men and one hundred heterosexual women between the ages of 18 and 40 rating each combination of face gender and trait (2000 raters in total). Trial order was fully randomised for each rater and the rating tasks were self paced.
}
\examples{
head(raw_3dsk_info)
}
\keyword{datasets}
