% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_stim.R
\name{load_stim}
\alias{load_stim}
\alias{load_stim_canada}
\alias{load_stim_london}
\alias{load_stim_zoom}
\alias{load_stim_smiling}
\alias{load_stim_composite}
\alias{load_stim_lisa}
\alias{load_stim_rainbow}
\title{Load image set}
\usage{
load_stim(
  set = c("london", "smiling", "composite", "canada", "lisa", "zoom", "rainbow"),
  pattern = NULL
)

load_stim_canada(pattern = NULL)

load_stim_london(pattern = NULL)

load_stim_zoom(pattern = NULL)

load_stim_smiling(pattern = NULL)

load_stim_composite(pattern = NULL)

load_stim_lisa(pattern = NULL)

load_stim_rainbow(pattern = NULL)
}
\arguments{
\item{set}{name of the image set}

\item{pattern}{Vector of patterns to use to search for files, or a vector of image indices}
}
\value{
a list with class stimlist
}
\description{
Load image sets as a stimlist, ready for processing with webmorphR functions.
}
\examples{
# loads the first 5 images in the set
stim <- load_stim("rainbow", 1:5)

# loads all images with 3 digits and m
canada_male <- load_stim("canada", "[0-9]{3}m")
}
